+++ .gitignore
@@ -13,17 +13,3 @@ Release/*
 .vs/*
 build/*
 .vscode/*
-**/CMakeCache.txt
-**/CMakeFiles/
-**/CMakeScripts/
-**/cmake_install.cmake
-**/Makefile
-.idea/
-cmake-build-*/
-**/CTestTestfile.cmake
-**/Testing/
-*.o
-*.a
-/delete_cache.cmake
-/Lab1CPP/Lab1CPP
-/Lab1CPP/output.txt
