+++ Lab1CPP/CMakeLists.txt
@@ -1,4 +1,12 @@
-add_executable(Lab1CPP lab1.cpp)
+add_executable(Lab1CPP
+        main.cpp
+        helpers.h
+        helpers.cpp
+        task1.h
+        task1.cpp
+        task2.cpp
+        task2.h
+)
 target_include_directories(Lab1CPP PUBLIC ../LibraryCPP)
 target_link_libraries(Lab1CPP LibraryCPP)
