+++ Lab1CPP/task1.h
@@ -1,0 +1,9 @@
+#ifndef ALGORITHMS_TASK1_H
+#define ALGORITHMS_TASK1_H
+
+#include "array.h"
+#include "helpers.h"
+
+void task1(Array *arr);
+
+#endif //ALGORITHMS_TASK1_H
