+++ Lab1CPP/task2.h
@@ -1,0 +1,9 @@
+#ifndef ALGORITHMS_TASK2_H
+#define ALGORITHMS_TASK2_H
+
+#include "array.h"
+#include "helpers.h"
+
+void task2(Array *arr);
+
+#endif //ALGORITHMS_TASK2_H
